requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==4.1.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
importlib-metadata==6.7.0
Markdown==3.4.4
mysqlclient==2.1.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
